Scanner

See merge request if3210-2024-android-pdd/if3210-2024-android-pdd!17

+++ app/src/main/java/com/example/bondoman_pdd/data/model/Scanner.kt
@@ -11,5 +11,5 @@ data class Items(
 data class ItemsInfo(
     val name: String,
     val qty: Int,
-    val price: Double
+    val price: Float
 )
